added gating to request resource
