+++ client/src/components/resources/ContactSubmitter.js
@@ -1,12 +1,7 @@
 import React from 'react'
-import { Anchor, Box, Text, Stack } from 'grommet'
-import { CreateOrLogin } from 'components/CreateAccountLoginButton'
-import styled from 'styled-components'
+import { Anchor, Box, Text } from 'grommet'
 import { useUser } from 'hooks/useUser'
-
-const LoginBox = styled(Box)`
-  backdrop-filter: blur(5px);
-`
+import { GatedLoginModal } from 'components/modals/GatedLoginModal'
 
 export const ContactSubmitter = ({ resource }) => {
   // check if logged in
@@ -20,7 +15,7 @@ export const ContactSubmitter = ({ resource }) => {
       <Text>
         Contact {contactUser.full_name} with questions about the resource.
       </Text>
-      <Stack guidingChild="last">
+      <GatedLoginModal title="Please sign in to contact submitter">
         <Box
           margin={{ vertical: 'medium', left: isLoggedIn ? 'none' : 'small' }}
         >
@@ -30,22 +25,7 @@ export const ContactSubmitter = ({ resource }) => {
             href={`mailto:${contactUser.email}`}
           />
         </Box>
-        {!isLoggedIn && (
-          <LoginBox align="center" pad="medium" full>
-            <Box
-              border={[{ color: 'black-tint-95' }]}
-              background="white"
-              elevation="medium"
-              pad="xlarge"
-            >
-              <CreateOrLogin
-                title="Please sign in to contact submitter"
-                showSignIn
-              />
-            </Box>
-          </LoginBox>
-        )}
-      </Stack>
+      </GatedLoginModal>
     </Box>
   )
 }
